# internal/api/rest/docs/openapi.yaml
openapi: 3.0.3
info:
  title: OpenTrader API
  description: |
    Enterprise-grade AI Trading Framework API
    
    ## Features
    - Multi-exchange trading (Crypto, Stocks, Forex, Options)
    - AI-powered market analysis
    - Strategy management
    - Real-time market data
    - Backtesting and optimization
  version: 1.0.0
  contact:
    name: OpenTrader Support
    url: https://github.com/opentreder/opentreder
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.opentreder.io/v1
    description: Production server
...
paths:
...
  /v1/orders:
    get:
      summary: List orders
      tags: [Orders]
      security:
        - BearerAuth: []
      parameters:
        - name: symbol
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [open, filled, cancelled, pending]
        - name: exchange
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: List of orders
          content:
            application/json:
              schema:
                type: object
                properties:
                  orders:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      summary: Place new order
      tags: [Orders]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlaceOrderRequest'
      responses:
        '201':
          description: Order placed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaceOrderResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/orders/{orderId}:
    get:
      summary: Get order details
      tags: [Orders]
      security:
        - BearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
        - name: exchange
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Order details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      summary: Cancel order
      tags: [Orders]
      security:
        - BearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
        - name: exchange
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Order cancelled
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/positions:
    get:
      summary: List all positions
      tags: [Positions]
      security:
        - BearerAuth: []
      parameters:
        - name: exchange
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of positions
          content:
            application/json:
              schema:
                type: object
                properties:
                  positions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Position'
  /v1/positions/{symbol}:
    get:
      summary: Get position for symbol
      tags: [Positions]
      security:
        - BearerAuth: []
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
        - name: exchange
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Position details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Position'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /v1/strategies:
    get:
      summary: List strategies
      tags: [Strategies]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of strategies
          content:
            application/json:
              schema:
                type: object
                properties:
                  strategies:
                    type: array
                    items:
                      $ref: '#/components/schemas/Strategy'

    post:
      summary: Create strategy
      tags: [Strategies]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStrategyRequest'
      responses:
        '201':
          description: Strategy created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Strategy'
  /v1/strategies/{strategyId}:
    get:
      summary: Get strategy
      tags: [Strategies]
      security:
        - BearerAuth: []
      parameters:
        - name: strategyId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Strategy details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Strategy'

    put:
      summary: Update strategy
      tags: [Strategies]
      security:
        - BearerAuth: []
      parameters:
        - name: strategyId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStrategyRequest'
      responses:
        '200':
          description: Strategy updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Strategy'

    delete:
      summary: Delete strategy
      tags: [Strategies]
      security:
        - BearerAuth: []
      parameters:
        - name: strategyId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Strategy deleted
  /v1/strategies/{strategyId}/start:
    post:
      summary: Start strategy
      tags: [Strategies]
      security:
        - BearerAuth: []
      parameters:
        - name: strategyId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Strategy started
  /v1/strategies/{strategyId}/stop:
    post:
      summary: Stop strategy
      tags: [Strategies]
      security:
        - BearerAuth: []
      parameters:
        - name: strategyId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Strategy stopped
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Order:
      type: object
      properties:
        order_id:
          type: string
        symbol:
          type: string
        exchange:
          type: string
        side:
          type: string
          enum: [buy, sell]
        order_type:
          type: string
          enum: [market, limit, stop, stop_limit, trailing_stop]
        status:
          type: string
          enum: [pending, open, filled, partially_filled, cancelled, expired]
        quantity:
          type: string
        filled_quantity:
          type: string
        price:
          type: string
        stop_price:
          type: string
        avg_fill_price:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    PlaceOrderRequest:
      type: object
      required:
        - symbol
        - side
        - quantity
      properties:
        symbol:
          type: string
        side:
          type: string
          enum: [buy, sell]
        order_type:
          type: string
          enum: [market, limit, stop, stop_limit]
          default: market
        quantity:
          type: string
        price:
          type: string
        stop_price:
          type: string
        time_in_force:
          type: string
          enum: [DAY, GTC, IOC, FOK]
          default: GTC
        exchange:
          type: string
    PlaceOrderResponse:
      type: object
      properties:
        order_id:
          type: string
        status:
          type: string
        message:
          type: string
    Position:
      type: object
      properties:
        symbol:
          type: string
        exchange:
          type: string
        quantity:
          type: string
        avg_entry_price:
          type: string
        current_price:
          type: string
        unrealized_pnl:
          type: string
        realized_pnl:
          type: string
        side:
          type: string
          enum: [long, short]
...
    Strategy:
      type: object
      properties:
        strategy_id:
          type: string
        name:
          type: string
        type:
          type: string
        parameters:
          type: object
          additionalProperties:
            type: string
        symbols:
          type: array
          items:
            type: string
        exchange:
          type: string
        enabled:
          type: boolean
        status:
          type: string
        created_at:
          type: string
          format: date-time
    CreateStrategyRequest:
      type: object
      required:
        - name
        - type
      properties:
        name:
          type: string
        type:
          type: string
          enum: [grid, dca, trend, scalping, market_making, mean_reversion, breakout, momentum, pairs]
        parameters:
          type: object
        symbols:
          type: array
          items:
            type: string
        exchange:
          type: string
        auto_start:
          type: boolean
          default: false
    UpdateStrategyRequest:
      type: object
      properties:
        parameters:
          type: object
        enabled:
          type: boolean
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              message:
                type: string
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              message:
                type: string